refactor: switch worktrees to be a map keyed by int
That way the interface of working with `worktrees` stays the same but
we get the ability of randomly delete elements from it. It basically
mirrors the interface of model.selected.

That made it possible to decouble deleting sideeffect (running git
commands) from deleting from the model. Which in its turn simplified
the logic for viewing the data and cursor fixing.